day11/requirements.txt
mcp>=1.2.0
langchain-groq
python-dotenv